+++ Chat Client Android Studio/app/src/main/java/com/milo/chatclientproject0315022020/ServerFeed.kt
@@ -41,7 +41,6 @@ class ServerFeed(private var context: Activity) : Runnable {
                         socket.close()
                         reader.close()
                         writer.close()
-                        Log.i("SERVER", "CLIENT CLOSED")
                     }
 
                     serverFeed(context, message)