+++ Chat Client Android Studio/app/src/main/java/com/milo/chatclientproject0315022020/ServerWrite.kt
@@ -23,7 +23,6 @@ class ServerWrite(private var message: String) : Runnable {
                 ChatMessage.serializer(),
                 ChatMessage(message, "userName", message, "createdDateAtTime")
             )
-
         writer.write((messageAsJson + '\n').toByteArray(Charset.defaultCharset()))
     }
 }