+++ Chat Server IntelliJ/src/main/kotlin/Utils.kt
@@ -1,6 +1,13 @@
 import java.text.SimpleDateFormat
 import java.util.*
 
+/**
+ * Class for utility functions.
+ *
+ * @author Michael Lock
+ * @date 11.02.2020
+ */
+
 object Utils {
 
     fun getCurrentTime(): String {