+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [0.40] - 29.02.2020
 
-### Added
+### Added features
 
 #### Chat Client
 - Connect screen has choice for choosing IP address and port
@@ -8,5 +8,8 @@
 - List history
 - List top users
 - Tap button for chat commands
+
+- Observable and singleton pattern
+